better func signatures

+++ streamer/util.go
@@ -13,7 +13,7 @@ const (
 	tb = 1000 * gb
 )
 
-func lexicalToNumber(size string) (int, error) {
+func lexicalToNumber(size string) (int64, error) {
 	// this condition is necessary for including infinite file size
 	if size == "-1" {
 		return -1, nil
@@ -37,5 +37,5 @@ func lexicalToNumber(size string) (int, error) {
 	if err != nil {
 		return 0, fmt.Errorf("couldn't parse input '%s' to float", size[:l-2])
 	}
-	return int(s * m), nil
+	return int64(s * m), nil
 }
